Refactored support for non-multisig newsrooms in core
The only fudge needed to make this happen was to add optional `owner`
property to `Contract` interface. This way both 1) a newsroom contract
can be sent to multisig proxy whether it's multisig or not and the proxy
will figure it out, and 2) non-ownable contracts like `CivilTCR` can be
sent to multisig proxy with or without multisig enabled.