Newsroom signup GA integration [CIVIL-477]
This tracks step movement as pageviews, and has events for newly
reaching a further step, as well as grant apply/skip and contract
creation TXs.

Events and actions are set up for tracking other TXs (tokens, applying)
and charter publish, but adding them all in was dragging on so I will do
the rest after launch.