Fix buggy logic for when to skip grant step
Buggy logic was "skip the grant step if the user is not waiting for
grant approval", but new users who haven't even selected whether they
want to apply for grant approval or not are of course "not waiting for
grant approval" so we erroneously skipped that step for them.

Similar logic is (correctly) used elsewhere so I've codified this into a
dedicated variable `completedGrantFlow` to make it clearer.