fix: use sentry in boost loader [CIVIL-1434]
The Story Boost embed itself loads the whole dapp wrapper in an iframe
and thus loads Sentry. However, the (relatively) minimal JS code that we
actually load on a newsroom's page that creates the ExternalLink post
and loads the iframe, that did not have Sentry.

Errors in the embed JS are important and need logging.

Adding Sentry brings the embed JS from 409kb (115kb gzipped) to 469kb
(131kb gzipped) so it's not great but not terrible.